fz-ehe.1: remove redundant as u64 casts at telemetry callsites — From<u32>/From<usize> handle conversion

+++ src/compiler2/drive.rs
@@ -97,7 +97,7 @@ impl World<'_> {
         let mut span = self.tel().span(
             &["fz", "compiler2", "drive"],
             metadata! {
-                pending_jobs: self.work_graph.pending_jobs() as u64,
+                pending_jobs: self.work_graph.pending_jobs(),
             },
         );
         let mut jobs_ran = 0_u64;
